requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest
